jsdoc and Alert moved to common components
